requirements.txt
Flask==3.0.3
azure-ai-textanalytics==5.3.0
azure-core==1.30.1
python-dotenv==1.0.1
